+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT UP
 
 - HP will no longer set max, will generate effects for hp bonuses like Toughness and Draconic Resilience when using effects.
+- Some features with double ''s in them would not get an icon.
 
 # 3.4.77