+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 - HP will no longer set max, will generate effects for hp bonuses like Toughness and Draconic Resilience when using effects.
 - Some features with double ''s in them would not get an icon.
-- MOre improvements to feature parsing in monsters.
+- More improvements to feature parsing in monsters.
+- Fixes for Planar Warrior Macro
 
 # 3.4.77